Moved psalm-refactor to the CLI command class
